+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Given a version number MAJOR.MINOR.PATCH, increment the:
 
 
 ## [Unreleased]
+### Changed
+- Updated renovate base branch 
+
 
 ## [v1.1.0] - 2022-10-26
 ### Added